+++ projects/ncv-example/src/app/app.component.html
@@ -7,9 +7,8 @@
   ></option>
 </select>
 
-<div class="container">
-  <nc-viewport
-    [toolList]="toolList"
-    [imageInfo]="imageInfos[imageIndex]"
-  ></nc-viewport>
-</div>
+<nc-viewport
+  class="container"
+  [toolList]="toolList"
+  [imageInfo]="imageInfos[imageIndex]"
+></nc-viewport>
