+++ projects/ncv-example/src/app/app.component.less
@@ -1,3 +1,8 @@
 select {
   min-width: 60px;
 }
+
+.container {
+  width: 100%;
+  height: 400px;
+}
