+++ projects/ng-cornerstone/package.json
@@ -2,13 +2,16 @@
   "name": "ng-cornerstone",
   "version": "0.0.1",
   "peerDependencies": {
-    "@angular/common": "^13.2.0",
-    "@angular/core": "^13.2.0",
+    "@angular/common": "^14.0.0",
+    "@angular/core": "^14.0.0",
     "@cornerstonejs/core": "^0.42.2",
     "@cornerstonejs/streaming-image-volume-loader": "^0.16.8",
-    "@cornerstonejs/tools": "^0.61.9"
+    "@cornerstonejs/tools": "^0.61.9",
+    "dcmjs": "^0.24.4",
+    "dicomweb-client": "^0.8.4"
   },
   "dependencies": {
+    "@angular/cdk": "^14.0.4",
     "tslib": "^2.3.0"
   }
 }
