+++ projects/ng-cornerstone/src/lib/components/button/button.component.html
@@ -1,13 +1,13 @@
 <ng-container [ngSwitch]="content">
   <ng-container *ngSwitchCase="'icon-text'">
-    <i [nc-icon]="icon"></i>
+    <i nc-icon [iconfont]="icon"></i>
     <span class="text">{{ text }}</span>
   </ng-container>
   <div *ngSwitchCase="'text'">
     <span>{{ text }}</span>
   </div>
   <div *ngSwitchCase="'icon'">
-    <i [nc-icon]="icon"></i>
+    <i nc-icon [iconfont]="icon"></i>
   </div>
   <ng-content *ngSwitchDefault></ng-content>
 </ng-container>
