+++ projects/ng-cornerstone/src/lib/core/init/initCornerstoneDICOMImageLoader.ts
@@ -1,10 +1,7 @@
 import dicomParser from 'dicom-parser';
 import * as cornerstone from '@cornerstonejs/core';
-import * as cornerstoneTools from '@cornerstonejs/tools';
 import cornerstoneWADOImageLoader from 'cornerstone-wado-image-loader';
 
-window.cornerstone = cornerstone;
-window.cornerstoneTools = cornerstoneTools;
 const { preferSizeOverAccuracy, useNorm16Texture } =
   cornerstone.getConfiguration().rendering;
