+++ projects/ng-cornerstone/src/lib/core/init/initProviders.ts
