--- projects/ng-cornerstone/src/lib/tool/tool-bar.component.scss
@@ -1,4 +1,0 @@
-:host {
-  display: inline-flex;
-  background-color: black;
-}
