+++ projects/ng-cornerstone/src/lib/viewport/viewport.component.html
@@ -1,4 +1,5 @@
 <nc-tool-bar
+  class="tool"
   [toolList]="toolList"
   [viewportType]="viewportType"
   [viewportId]="viewportId"
