+++ projects/ng-cornerstone/src/public-api.ts
@@ -2,13 +2,15 @@
  * Public API Surface of ng-cornerstone
  */
 export * from './lib/core/config/types';
-export * from './lib/viewport/viewport.component';
-export * from './lib/viewport/viewport.module';
 export * from './lib/components/button/button.component';
 export * from './lib/components/button/button-group.component';
 export * from './lib/components/button/button.module';
 export * from './lib/components/icon/icon.component';
 export * from './lib/components/icon/icon.module';
+// export * from './lib/tool';
+export * from './lib/tool/tool.types';
 export * from './lib/tool/tool.config';
 export * from './lib/tool/tool-bar.component';
 export * from './lib/tool/tool.module';
+export * from './lib/viewport/viewport.component';
+export * from './lib/viewport/viewport.component';
